Simplify module structure
